fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
click==8.3.0
fastapi==0.117.1
gitdb==4.0.12
GitPython==3.1.41
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pystache==0.6.8
python-multipart==0.0.20
setuptools==75.6.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
